% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_check_data.r
\name{wat_check_data}
\alias{wat_check_data}
\title{Check data has required columns.}
\usage{
wat_check_data(data, names_expected = c("x", "y", "time"))
}
\arguments{
\item{data}{The tracking data to check for required columns.}

\item{names_expected}{The names expected.}
}
\value{
Nothing. Breaks a pipeline if the asserts fail.
}
\description{
Check data has required columns.
}
