% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fix_speed.r
\name{wat_fix_speeds}
\alias{wat_fix_speeds}
\title{Fix extreme speeds and reflections.}
\usage{
wat_fix_speeds(data, x = "x", y = "y", time = "time", speed_cutoff)
}
\arguments{
\item{data}{A dataframe or similar, containing the columns \code{x}, \code{y}, and \code{time}.}

\item{speed_cutoff}{A speed cutoff in metres per second.}
}
\value{

}
\description{
Fix extreme speeds and reflections.
}
