% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_angle.r
\name{wat_get_angle}
\alias{wat_get_angle}
\title{Get the angle (relative bearing) between points.}
\usage{
wat_get_angle(data, x = "x", y = "y", time = "time")
}
\arguments{
\item{data}{A dataframe or similar which must have the columns
specified by \code{x}, \code{y}, and \code{time}.}

\item{x}{The x coordinate.}

\item{y}{The y coordinate.}

\item{time}{The timestamp in seconds since the UNIX epoch.}
}
\value{
A vector of relative bearings in degrees. Negative degrees indicate 'left' turns.
}
\description{
Get the angle (relative bearing) between points.
}
