% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_getData.r
\name{wat_get_data}
\alias{wat_get_data}
\title{Get data from a remote ATLAS SQL database.}
\usage{
wat_get_data(
  tag,
  tracking_time_start,
  tracking_time_end,
  database = "some_database",
  host = "abtdb1.nioz.nl",
  username = "someuser",
  password = "somepassword"
)
}
\arguments{
\item{tag}{An 11 digit number representing the ATLAS tag.
May be passed as a character or a numeric.
Will be converted to a character.
An example is \code{"31001000001"}.}

\item{tracking_time_start}{Character representation of time
(Central European Time, +1 UTC) from which start point data
should be retrieved.}

\item{tracking_time_end}{Character time representing the end point
corresponding to the above start point.}

\item{database}{The database name on the host server. Defaults to
\code{db} for unknown reasons.}

\item{host}{The server address on which the data are stored.
Defaults to \code{abtdb1.nioz.nl}.}

\item{username}{Username to access the data.}

\item{password}{Password to access the data.}
}
\value{
A dataframe of agent positions, filtered between the
required tracking times.
}
\description{
Get data from a remote ATLAS SQL database.
}
