% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_count_reflections.r
\name{wat_count_reflections}
\alias{wat_count_reflections}
\title{Count data reflections.}
\usage{
wat_count_reflections(
  data,
  x = "x",
  y = "y",
  time = "time",
  speed_cutoff = 0.5
)
}
\arguments{
\item{data}{The data in which to count so-called reflections.}

\item{speed_cutoff}{The speed cutoff in metres per second by which to
determine whether positions are reflected or not.}
}
\value{
A named list with two elements \code{n_reflections} and
\code{p_reflected}, the first indicating how many times there are
reflecions in the data, and the second indicating the proportion of fixes
that have been reflected.
}
\description{
Count data reflections.
}
