% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fix_speed.r
\name{wat_fix_speeds}
\alias{wat_fix_speeds}
\title{Fix extreme speeds and reflections.}
\usage{
wat_fix_speeds(data, x = "x", y = "y", time = "time", speed_cutoff)
}
\arguments{
\item{data}{A dataframe or similar, containing the columns \code{x},
\code{y}, and \code{time}.}

\item{x}{The X coordinate.}

\item{y}{The Y coordinate.}

\item{time}{The timestamp, preferably in seconds since the UNIX epoch.}

\item{speed_cutoff}{A speed cutoff in metres per second, above which to
remove both point locations as well as reflections.}
}
\value{
A dataframe with extreme speeds and reflections removed.
}
\description{
Fix extreme speeds and reflections.
}
